feat: 🎸 Adds indeterminate state and separes checkbox
